files/requirements.txt
flask>=3.0.0
python-dotenv>=1.0.0
google-genai>=0.8.0
scikit-learn>=1.4.0
numpy>=1.26.0
